:recycle: refactor : user 타입 정의, Sidebar 분리
